Split up biome types, move interfaces to api-package
